requirements.txt
numpy
matplotlib
wandb
python-dotenv